Move `TxIn` type into separate module.
This allows the `TxIn` type to be imported without having to depend on
everything in the `Tx.Tx` module.
